move type declarations to types.ts
