+++ src/commands/dev/dev-router.ts
@@ -2,43 +2,11 @@ import http from 'http'
 import url from 'url'
 import qs from 'querystring'
 
-// @ts-ignore
-import FileFsRef from '@now/build-utils/file-fs-ref';
-// @ts-ignore
-import Lambda from '@now/build-utils/lambda'
-
-type BuilderOutput = FileFsRef | Lambda
-
-interface BuilderOutputs {
-  [key: string]: BuilderOutput
-}
-
-interface HttpHeaderConfig {
-  [header: string]: string
-}
-
-interface RouteConfig {
-  src: string,
-  dest: string,
-  methods?: string[],
-  headers?: HttpHeaderConfig[],
-  status?: number
-}
-
-interface RouteResult {
-  // "dest": <string of the dest, either file for lambda or full url for remote>
-  dest: BuilderOutput | string,
-  // "status": <integer in case exit code is intended to be changed>
-  status?: number,
-  // "headers": <object of the added response header values>
-  headers?: HttpHeaderConfig[],
-  // "uri_args": <object (key=value) list of new uri args to be passed along to dest >
-  uri_args?: {[key: string]: any},
-  // "matched_route": <object of the route spec that matched>
-  matched_route?: RouteConfig,
-  // "matched_route_idx": <integer of the index of the route matched>
-  matched_route_idx?: number
-}
+import {
+  BuilderOutputs,
+  RouteConfig,
+  RouteResult
+} from './types';
 
 export default function (
   req: http.IncomingMessage,
